The Society Management project (Open Source API Management)
The Society Management project is a standalone API Management system that can be either run as a separate system or embedded within existing frameworks and platforms.

The above maven command will do the following:

A full build of Society Management.
Download WildFly 10.x
Unpack and configure WildFly 10.x
Deploy the WildFly 10.x version of Society Management to WildFly
Once WildFly has started up, and if all went well, you can point your browser to the API Manager and log in (either register a new user or log in as the admin):

admin/admin123!
Note that this quickstart seeds a bunch of content into Society Management. This is not strictly necessary, but it does populate the API Manager with some data so that it doesn't feel so lonely the first time you log in. You're welcome.

After you're happy with your changes and a full build (with unit tests) runs successfully, commit your changes on your topic branch. Then it's time to check for and pull any recent changes that were made in the official repository:

We prefer pull-requests because we can review the proposed changes, comment on them, discuss them with you, and likely merge the changes right into the official repository.

Please try to create one commit per feature or fix, generally, the easiest way to do this is via git squash. This makes reverting changes easier, and avoids needlessly polluting the repository history with checkpoint commits.